Construct a Repository for Katas
Using singleton classes, a repository for katas can be constructed. The
model for a Kata should feature the following attributes:
 - 'title' for the title of the kata
 - 'link' for the URL of the video used in the kata
 - 'description' for details about the kata

The Repository itself should be accessible through a singleton
attribute. A base 'Repository' class should refer to one and only one
instance of the 'KataRepository' class through a class method ::kata.