Create service for handling all media
The MediaService static class should handle routing generic requests to
the appropriate services.

The :embed static method should receive a URI instance and return the
result of passing that URI to the appropriate service method.

In the case of URIs with a host of 'vimeo.com', the result of
VimeoService.embed should be returned.